+++ common.h
@@ -11,7 +11,7 @@
 // @param d0 first d in equation
 // @param i1 second i in equation
 // @param d1 second d in equation
-std::vector<int> interpolate(int i0, float d0, int i1, float d1);
+std::vector<float> interpolate(int i0, float d0, int i1, float d1);
 
 // wrapper function for drawing functions(translates the pixels)
 // uses the logical coordinate system(-screen_width/2 to screen_width/2, etc.)
